Fix handling of Openshift memtotal and change datatype to Long

+++ src/main/java/org/candlepin/insights/inventory/ConduitFacts.java
@@ -77,7 +77,7 @@ public class ConduitFacts extends ConsumerInventory {
 
     @Positive
     @Override
-    public Integer getMemory() {
+    public Long getMemory() {
         return super.getMemory();
     }
 }
